+++ src/MetricsTest.php
@@ -20,18 +20,8 @@ final class MetricsTest extends TestCase
      */
     private function createSummary(string $label): Summary
     {
-        $start = new Snapshot('start', new DateTimeImmutable(), hrtime(true), [
-            'ru_utime.tv_sec' => 1,
-            'ru_stime.tv_sec' => 1,
-            'ru_utime.tv_usec' => 1,
-            'ru_stime.tv_usec' => 1,
-        ], 1000, 2000, 3000, 4000);
-        $end = new Snapshot('end', new DateTimeImmutable(), hrtime(true) + 1, [
-            'ru_utime.tv_sec' => 1,
-            'ru_stime.tv_sec' => 1,
-            'ru_utime.tv_usec' => 1,
-            'ru_stime.tv_usec' => 1,
-        ], 1100, 2100, 3100, 4100);
+        $start = new Snapshot('start', new DateTimeImmutable(), hrtime(true), 10001, 1001, 1000, 2000, 3000, 4000);
+        $end = new Snapshot('end', new DateTimeImmutable(), hrtime(true) + 1, 1001, 1001, 1100, 2100, 3100, 4100);
         return new Summary($label, $start, $end);
     }
 
@@ -71,42 +61,62 @@ final class MetricsTest extends TestCase
     #[Test]
     public function it_correctly_returns_the_cpu_time_in_nanoseconds(): void
     {
-        $start = new Snapshot('start', new DateTimeImmutable(), hrtime(true), [
-            'ru_utime.tv_sec'  => 1,
-            'ru_utime.tv_usec' => 500_000, // 1.5s user
-            'ru_stime.tv_sec'  => 0,
-            'ru_stime.tv_usec' => 250_000, // 0.25s system
-        ], 1000, 2000, 3000, 4000);
-        $end = new Snapshot('end', new DateTimeImmutable(), hrtime(true) + 1, [
-            'ru_utime.tv_sec'  => 2,
-            'ru_utime.tv_usec' => 0,       // 2.0s user
-            'ru_stime.tv_sec'  => 0,
-            'ru_stime.tv_usec' => 750_000, // 0.75s system
-        ], 1100, 2100, 3100, 4100);
+        $start = new Snapshot(
+            'start',
+            new DateTimeImmutable(),
+            hrtime(true),
+            1_500_000,
+            250_000,
+            1000,
+            2000,
+            3000,
+            4000,
+        );
+        $end = new Snapshot(
+            'end',
+            new DateTimeImmutable(),
+            hrtime(true) + 1,
+            2_000_000,
+            750_000,
+            1100,
+            2100,
+            3100,
+            4100,
+        );
 
         // Expected:
         // User time: 2.0 - 1.5 = 0.5s = 500_000_000 ns
         // System time: 0.75 - 0.25 = 0.5s = 500_000_000 ns
         // Total = 1_000_000_000 ns
 
-        self::assertSame(1_000_000_000.0, Metrics::fromSnapshots($start, $end)->cpuTime);
+        self::assertSame(1_000_000.0, Metrics::fromSnapshots($start, $end)->cpuTime);
     }
 
     #[Test]
     public function it_can_returns_the_metrics_in_a_human_readable_format(): void
     {
-        $start = new Snapshot('start', new DateTimeImmutable(), hrtime(true), [
-            'ru_utime.tv_sec'  => 1,
-            'ru_utime.tv_usec' => 500_000, // 1.5s user
-            'ru_stime.tv_sec'  => 0,
-            'ru_stime.tv_usec' => 250_000, // 0.25s system
-        ], 1000, 2000, 3000, 4000);
-        $end = new Snapshot('end', new DateTimeImmutable(), hrtime(true) + 1, [
-            'ru_utime.tv_sec'  => 2,
-            'ru_utime.tv_usec' => 0,       // 2.0s user
-            'ru_stime.tv_sec'  => 0,
-            'ru_stime.tv_usec' => 750_000, // 0.75s system
-        ], 1100, 2100, 3100, 4100);
+        $start = new Snapshot(
+            'start',
+            new DateTimeImmutable(),
+            hrtime(true),
+            1_500_000,
+            250_000,
+            1000,
+            2000,
+            3000,
+            4000,
+        );
+        $end = new Snapshot(
+            'end',
+            new DateTimeImmutable(),
+            hrtime(true) + 1,
+            2_000_000,
+            750_000,
+            1100,
+            2100,
+            3100,
+            4100,
+        );
         $metrics = Metrics::fromSnapshots($start, $end);
 
         $humans = $metrics->forHuman();